refactor: add CI & drop gulp & use TypeScript
