+++ .mocharc.json
@@ -1,0 +1,5 @@
+{
+  "require": [
+    "ts-node-test-register"
+  ]
+}
